Binded a dummy function to double-clicked entries.
The message "Do nothing" is now printed when a hashtag or a keyword from
a listbox is double-clicked. This is to be modified so as to open the
associated URL in a browser.

The ".gitignore" file was also expanded a bit to include some more files.

An unused library "PIL" was removed from the file "gui.py" as well.