Common Treeview updates internally + has url col.
Instead of creating a new Treeview element each time the common hashtags
and keywords are updated, the Treeview now updates its elements
internally by first deleting all the rows then inserting the new ones in
the original Treeview element.
Note: This feature was implemented before but replaced. The feature was revived in
this commit.

However, as the double-click binding has to be renewed which might
introduce problems, the system was changed so that the url is included
as an invisible column in the Treeview. This means that the url will
definitely point to the correct query.