Made openLink independent of the url col position.
Problem: Previously, openLink assumed that the url was in the 2nd col, which is a
dangerous assumption. Were the positions of the columns to be switched,
the function itself will have to be modified manually.

Fix: By referring to the fields list used to name the columns, the
values obtained from the selected row are first converted to a
dictionary. Thus, openLink now refers to the 'url' col rather than the
nth col, which is a better representative of the thing should actually
work and makes maintenance easier.